Refactored process handling
